+++ examples/haskell/flake.nix
@@ -34,13 +34,10 @@
                       fileName = lastSafe (lib.strings.splitString "/" path);
                     in
                     fileName != "flake.nix" &&
-                      fileName != "garn.ts";
+                    fileName != "garn.ts";
                 }
             )
-            { })
-          // {
-            meta.mainProgram = "helloFromHaskell";
-          };
+            { });
         }
       );
       checks = forAllSystems (system:
@@ -74,13 +71,10 @@
                               fileName = lastSafe (lib.strings.splitString "/" path);
                             in
                             fileName != "flake.nix" &&
-                              fileName != "garn.ts";
+                            fileName != "garn.ts";
                         }
                     )
-                    { })
-                  // {
-                    meta.mainProgram = "helloFromHaskell";
-                  };
+                    { });
                 in
                 (if expr ? env
                 then expr.env
@@ -104,8 +98,9 @@
               } "
     touch \$out
     ${"
-      echo copying source
-      cp -r ${(let
+      ${"
+    echo copying source
+    cp -r ${(let
     lib = pkgs.lib;
     lastSafe = list :
       if lib.lists.length list == 0
@@ -123,8 +118,10 @@
          fileName != "flake.nix" &&
          fileName != "garn.ts";
     })} src
-      chmod -R u+rwX src
-      cd src
+    chmod -R u+rwX src
+    cd src
+  "}
+      ${""}
     "}
     ${"hlint *.hs"}
   ";
@@ -159,13 +156,10 @@
                           fileName = lastSafe (lib.strings.splitString "/" path);
                         in
                         fileName != "flake.nix" &&
-                          fileName != "garn.ts";
+                        fileName != "garn.ts";
                     }
                 )
-                { })
-              // {
-                meta.mainProgram = "helloFromHaskell";
-              };
+                { });
             in
             (if expr ? env
             then expr.env
@@ -192,10 +186,35 @@
           };
         in
         {
-          "helloFromHaskell" = {
+          "helloFromHaskell/hello" = {
             "type" = "app";
             "program" = "${let
-        dev = pkgs.mkShell {};
+        dev = let expr = (pkgs.haskell.packages.ghc94.callCabal2nix
+        "garn-pkg"
+        (let
+    lib = pkgs.lib;
+    lastSafe = list :
+      if lib.lists.length list == 0
+        then null
+        else lib.lists.last list;
+  in
+  builtins.path
+    {
+      path = ./.;
+      name = "source";
+      filter = path: type:
+        let
+          fileName = lastSafe (lib.strings.splitString "/" path);
+        in
+         fileName != "flake.nix" &&
+         fileName != "garn.ts";
+    })
+        { });
+        in
+          (if expr ? env
+            then expr.env
+            else pkgs.mkShell { inputsFrom = [ expr ]; }
+          );
         shell = "${(pkgs.haskell.packages.ghc94.callCabal2nix
         "garn-pkg"
         (let
@@ -216,10 +235,7 @@
          fileName != "flake.nix" &&
          fileName != "garn.ts";
     })
-        { })
-        // {
-          meta.mainProgram = "helloFromHaskell";
-        }}/bin/${"helloFromHaskell"}";
+        { })}/bin/${"hello"}";
         buildPath = pkgs.runCommand "build-inputs-path" {
           inherit (dev) buildInputs nativeBuildInputs;
         } "echo $PATH > $out";