+++ test/spec/BuildSpec.hs
@@ -114,10 +114,12 @@ spec = do
             [i|
               import * as garn from "#{repoDir}/ts/mod.ts"
 
-              const env = garn.mkEnvironment(undefined, garn.nix.nixStrLit`
-                mkdir dist
-                echo build-content > dist/build-artifact
-              `);
+              const env = garn.mkEnvironment({
+                setup: garn.nix.nixStrLit`
+                  mkdir dist
+                  echo build-content > dist/build-artifact
+                `
+              });
               export const project = garn.mkProject({
                 description: "",
                 defaultEnvironment: env,
@@ -179,10 +181,9 @@ spec = do
                 { description: "" },
                 {
                   package: garn
-                    .mkEnvironment(
-                      undefined,
-                      nix.nixStrLit`SETUP_VAR="hello from setup"`,
-                    )
+                    .mkEnvironment({
+                      setup: nix.nixStrLit`SETUP_VAR="hello from setup"`,
+                    })
                     .build("echo $SETUP_VAR > $out/build-artifact"),
                 },
               )
@@ -198,3 +199,19 @@ spec = do
         doesDirectoryExist "result" `shouldReturn` True
         StdoutTrim output <- cmd ("result/bin/garn-test" :: String)
         output `shouldBe` ("haskell test output" :: String)
+
+      it "does not warn about dirty git directories" $ \runGarn -> do
+        cmd_ "git init --initial-branch=main" (EchoStdout False)
+        writeFile "file" "foo"
+        writeFile
+          "garn.ts"
+          [i|
+            import * as garn from "#{repoDir}/ts/mod.ts";
+
+            export const pkg = garn.build("echo built > $out/artifact");
+          |]
+        cmd_ "git add ."
+        cmd_ "git commit -m" ["test commit message"] (EchoStdout False)
+        writeFile "file" "bar"
+        output <- runGarn ["build", "pkg"]
+        stderr output `shouldNotMatch` "Git tree .* is dirty"