+++ ts/project.ts
@@ -189,7 +189,7 @@ export function isProject(p: unknown): p is Project {
   return hasTag(p, "project");
 }
 
-type Nestable = Environment | Package | Executable | Check | Project;
+export type Nestable = Environment | Package | Executable | Check | Project;
 
 /**
  * Creates a new `Project`.