+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 - Add `addCabalExecutable` to Haskell projects
 - Make `garn init` on Haskell projects fill in the `executables` based on the cabal file.
 
+- For git repositories, add untracked files to build sandboxes for `Check`s and
+  `Package`s.
+
+- Tweak the API for creating `Environment`s, making it easier to create
+  `Environment`s that have a source directory
+
 ## v0.0.16
 
 - Allow to build packages that are nested within projects with `garn build projectName.packageName`.